Added capability to run JUnit5 test with different dependencies
