+++ pom.xml
@@ -104,6 +104,7 @@
         <version.h2>1.4.196</version.h2>
         <version.jackson>[2.10.0,)</version.jackson>
         <version.junit-jupiter>5.9.1</version.junit-jupiter>
+        <version.junit-platform-launcher>1.9.1</version.junit-platform-launcher>
         <version.junit.vintage>4.13.2</version.junit.vintage>
         <version.junit-vintage-engine>5.9.1</version.junit-vintage-engine>
         <version.logback>1.2.3</version.logback>
@@ -633,6 +634,13 @@
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-engine</artifactId>
             <version>${version.junit-jupiter}</version>
+            <scope>test</scope>
+        </dependency>
+        <!-- platform launcher is required for programmatic launching of JUnit5 tests -->
+        <dependency>
+            <groupId>org.junit.platform</groupId>
+            <artifactId>junit-platform-launcher</artifactId>
+            <version>${version.junit-platform-launcher}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
